+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - `storage_integration`
     - `credentials`
     - `encryption`
+- Added support for `Session.directory` to retrieve the list of all files on a stage with metadata.
 
 #### Bug Fixes